Updated Adobe AIR 19 Added myo.framework dependency

+++ src/Preset.as
@@ -97,7 +97,7 @@ package
 		public function remove():void 
 		{
 			var f:File = getPresetFile();
-			f.deleteFile();
+			if(f.exists) f.deleteFile();
 		}
 		
 		//HANDLERS
